增加Command结构,思考初始化流程和方法，处理command解析 优化整理file的socket发送 完成master的上传和下载命令处理，file的socket发送 upload、download task clear 完成大部分初始化、管理命令处理 下载任务超时计数、上传下载处理 完成下载Store命令处理 store server上传命令处理 下一步： backup错处理。 上传MD5验证计算 其他命令处理：index、delete

+++ common/src/main/java/com/ly/common/domain/file/FileInfo.java
@@ -12,7 +12,7 @@ public class FileInfo {
     private String etag;
     private boolean chunk;
     // 分块大小，<=0表示不分块
-    private long chunkSize;
+    private int chunkSize;
     // 分块数量
     private int chunkCount;
 
@@ -75,11 +75,11 @@ public class FileInfo {
         this.chunk = chunk;
     }
 
-    public long getChunkSize() {
+    public int getChunkSize() {
         return chunkSize;
     }
 
-    public void setChunkSize(long chunkSize) {
+    public void setChunkSize(int chunkSize) {
         this.chunkSize = chunkSize;
     }
